functions/email_notifier/requirements.txt
azure-functions
azure-functions-durable
msal
requests
